+++ javascript/node/selenium-webdriver/example/google_search_test.js
@@ -18,40 +18,56 @@
 /**
  * @fileoverview An example test that may be run using Mocha.
  *
+ * This example uses the `selenium-webdriver/testing.suite` function, which will
+ * automatically run tests against every available WebDriver browser on the
+ * current system. Alternatively, you may use the `SELENIUM_BROWSER`
+ * environment variable to narrow the scope at runtime.
+ *
  * Usage:
  *
+ *     # Automatically determine which browsers to run against.
  *     mocha -t 10000 selenium-webdriver/example/google_search_test.js
  *
- * You can change which browser is started with the SELENIUM_BROWSER environment
- * variable:
- *
+ *     # Configure tests to only run against Google Chrome.
  *     SELENIUM_BROWSER=chrome \
  *         mocha -t 10000 selenium-webdriver/example/google_search_test.js
  */
 
-const {Builder, By, Key, until} = require('..');
+const assert = require('assert');
 
-describe('Google Search', function() {
-  let driver;
+const {Browser, By, Key, until} = require('..');
+const {ignore, suite} = require('../testing');
 
-  before(async function() {
-    driver = await new Builder().forBrowser('firefox').build();
+suite(function(env) {
+  describe('Google Search', function() {
+    let driver;
+
+    before(async function() {
+      // env.builder() returns a Builder instance preconfigured for the
+      // envrionment's target browser (you may still define browser specific
+      // options if necessary (i.e. firefox.Options or chrome.Options)).
+      driver = await env.builder().build();
+    });
+
+    it('demo', async function() {
+      await driver.get('https://www.google.com/ncr');
+      await driver.findElement(By.name('q')).sendKeys('webdriver', Key.RETURN);
+      await driver.wait(until.titleIs('webdriver - Google Search'), 1000);
+    });
+
+    // The ignore function returns wrappers around describe & it that will
+    // suppress tests if the provided predicate returns true. You may provide
+    // any synchronous predicate. The env.browsers(...) function generates a
+    // predicate that will suppress tests if the  env targets one of the
+    // specified browsers.
+    //
+    // This example is always configured to skip Chrome.
+    ignore(env.browsers(Browser.CHROME)).it('demo 2', async function() {
+      await driver.get('http://www.google.com/ncr');
+      let url = await driver.getCurrentUrl();
+      assert.equal(url, 'https://www.google.com/');
+    });
+
+    after(() => driver && driver.quit());
   });
-
-  // You can write tests either using traditional promises.
-  it('works with promises', function() {
-    return driver.get('http://www.google.com/ncr')
-        .then(_ =>
-            driver.findElement(By.name('q')).sendKeys('webdriver', Key.RETURN))
-        .then(_ => driver.wait(until.titleIs('webdriver - Google Search'), 1000));
-  });
-
-  // Or you can define the test with an async function.
-  it('works with generators', async function() {
-    await driver.get('http://www.google.com/ncr');
-    await driver.findElement(By.name('q')).sendKeys('webdriver', Key.RETURN);
-    await driver.wait(until.titleIs('webdriver - Google Search'), 1000);
-  });
-
-  after(() => driver.quit());
 });