+++ javascript/node/selenium-webdriver/ie.js
@@ -352,6 +352,18 @@ class Options {
 }
 
 
+/**
+ * _Synchronously_ attempts to locate the IE driver executable on the current
+ * system.
+ *
+ * @return {?string} the located executable, or `null`.
+ */
+function locateSynchronously() {
+  return process.platform === 'win32'
+      ? io.findInPath(IEDRIVER_EXE, true) : null;
+}
+
+
 function createServiceFromCapabilities(capabilities) {
   if (process.platform !== 'win32') {
     throw Error(
@@ -360,7 +372,7 @@ function createServiceFromCapabilities(capabilities) {
         'WebDriver server?');
   }
 
-  let exe = io.findInPath(IEDRIVER_EXE, true);
+  let exe = locateSynchronously();
   if (!exe || !fs.existsSync(exe)) {
     throw Error(
         `${IEDRIVER_EXE} could not be found on the current PATH. Please ` +
@@ -437,3 +449,5 @@ class Driver extends webdriver.WebDriver {
 exports.Driver = Driver;
 exports.Options = Options;
 exports.Level = Level;
+exports.locateSynchronously = locateSynchronously;
+